Personal Project website for education.
The idea was to use this material as a way to teach kids English by practicing prepositional phrases and basic word definition. "The rock goes between the blue and red fish", "The cloud is at the top right corner". A great interactive way for students to not only learn new words, but to practice  speaking using prepositions.